Add a spec for throwing GuildNotFound when get_guild is given an invalid guild
When given a non-existent guild, the Armory throws a 500 error. Prior to
this commit, Wowr would ignore the response and keep retrying the
request until the limit was reached, and then would throw a network
timeout error, which was incorrect.

This change handles the 500 response by returning the error response
body, which now works for get_guild but may break other methods. We'll
find those out through further testing.