Create an API namespace
In order to break the API class into multiple classes, we need an API
module. This moves the API class from the Wowr module to the Wowr::API
module. In order to not break old code using Wowr, we create a class
method on Wowr::API called 'new' which returns an instance of
Wowr::API::API.